refactor: extract and test pickPropsDeep generic types
